refactor: improve serivceConfigurationKey
